requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-async-sqlalchemy==0.6.0
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.4
install==1.3.5
markdown-it-py==3.0.0
mdurl==0.1.2
passlib==1.7.4
priority==2.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.16.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rich==13.7.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
wsproto==1.2.0
